test(nostr): consolidate metrics coverage at its owner
